feat: create package
Create peakoscope package with new namespace: peakoscope/__init__.py. Rename cli.py -> peakoscope/__main__.py to get CLI when running package.
Move all modules (.py) to peakoscope/. Modify imports in all modules (.py) and notebooks (.ipynb).
